move `worklets` directory to `workers`
